Change from int to Long
-Organize imports

+++ backend/src/main/java/ch/zhaw/parkship/parkinglot/ParkingLotDto.java
@@ -5,7 +5,6 @@ import java.util.HashSet;
 import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-
 import ch.zhaw.parkship.tag.TagDto;
 import ch.zhaw.parkship.user.UserDto;
 import jakarta.validation.constraints.NotBlank;
@@ -17,78 +16,78 @@ import lombok.Setter;
 @Setter
 public class ParkingLotDto implements Serializable {
 
-    private Integer id;
+  private Long id;
 
-    @NotNull
-    private UserDto owner;
+  @NotNull
+  private UserDto owner;
 
-    private String description;
+  private String description;
 
-    private Set<TagDto> tags;
+  private Set<TagDto> tags;
 
-    @NotNull
-    private Double longitude;
+  @NotNull
+  private Double longitude;
 
-    @NotNull
-    private Double latitude;
+  @NotNull
+  private Double latitude;
 
-    private String address;
+  private String address;
 
-    private String addressNr;
+  private String addressNr;
 
-    private Integer floor;
+  private Integer floor;
 
-    @NotBlank
-    private String nr;
+  @NotBlank
+  private String nr;
 
-    private byte[] picture;
+  private byte[] picture;
 
-    @NotNull
-    private Double price;
+  @NotNull
+  private Double price;
 
-    @NotBlank
-    private String state;
+  @NotBlank
+  private String state;
 
-    public ParkingLotDto(ParkingLotEntity parkingLotEntity) {
-        this.id = parkingLotEntity.getId();
-        this.owner = new UserDto(parkingLotEntity.getOwner());
-        this.description = parkingLotEntity.getDescription();
-        this.tags = parkingLotEntity.getTags().stream().map(TagDto::new).collect(Collectors.toSet());
-        this.longitude = parkingLotEntity.getLongitude();
-        this.latitude = parkingLotEntity.getLatitude();
-        this.address = parkingLotEntity.getAddress();
-        this.addressNr = parkingLotEntity.getAddressNr();
-        this.floor = parkingLotEntity.getFloor();
-        this.nr = parkingLotEntity.getNr();
-        this.picture = parkingLotEntity.getPicture();
-        this.price = parkingLotEntity.getPrice();
-        this.state = parkingLotEntity.getState();
-    }
+  public ParkingLotDto(ParkingLotEntity parkingLotEntity) {
+    this.id = parkingLotEntity.getId();
+    this.owner = new UserDto(parkingLotEntity.getOwner());
+    this.description = parkingLotEntity.getDescription();
+    this.tags = parkingLotEntity.getTags().stream().map(TagDto::new).collect(Collectors.toSet());
+    this.longitude = parkingLotEntity.getLongitude();
+    this.latitude = parkingLotEntity.getLatitude();
+    this.address = parkingLotEntity.getAddress();
+    this.addressNr = parkingLotEntity.getAddressNr();
+    this.floor = parkingLotEntity.getFloor();
+    this.nr = parkingLotEntity.getNr();
+    this.picture = parkingLotEntity.getPicture();
+    this.price = parkingLotEntity.getPrice();
+    this.state = parkingLotEntity.getState();
+  }
 
-    public ParkingLotDto() {
-        this.tags = new HashSet<>();
-    }
+  public ParkingLotDto() {
+    this.tags = new HashSet<>();
+  }
 
-    @Override
-    public String toString() {
-        return "ParkingLotDto{" + "id=" + id + ", owner=" + owner + ", description='" + description + '\'' + ", tags="
-                + tags + ", longitude=" + longitude + ", latitude=" + latitude + ", address='" + address + '\''
-                + ", addressNr='" + addressNr + '\'' + ", floor=" + floor + ", nr='" + nr + '\'' + ", price=" + price
-                + ", state='" + state + '\'' + '}';
-    }
+  @Override
+  public String toString() {
+    return "ParkingLotDto{" + "id=" + id + ", owner=" + owner + ", description='" + description
+        + '\'' + ", tags=" + tags + ", longitude=" + longitude + ", latitude=" + latitude
+        + ", address='" + address + '\'' + ", addressNr='" + addressNr + '\'' + ", floor=" + floor
+        + ", nr='" + nr + '\'' + ", price=" + price + ", state='" + state + '\'' + '}';
+  }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
+  @Override
+  public int hashCode() {
+    return Objects.hash(id);
+  }
 
-    @Override
-    public boolean equals(Object o) {
-        if (this == o)
-            return true;
-        if (!(o instanceof ParkingLotDto))
-            return false;
-        ParkingLotDto that = (ParkingLotDto) o;
-        return Objects.equals(id, that.id);
-    }
+  @Override
+  public boolean equals(Object o) {
+    if (this == o)
+      return true;
+    if (!(o instanceof ParkingLotDto))
+      return false;
+    ParkingLotDto that = (ParkingLotDto) o;
+    return Objects.equals(id, that.id);
+  }
 }
